requirements.txt
asgiref==3.4.1
astroid==2.9.3
autopep8==1.6.0
colorama==0.4.4
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pycodestyle==2.8.0
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
wrapt==1.13.3
